refactor(compute): split backends into modules
