+++ crates/compute/src/backend/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "mock")]
+pub mod mock_cpu;
+
+#[cfg(all(target_os = "macos", feature = "metal"))]
+pub mod wgpu_metal;
