+++ crates/compute/src/kernels/expand_instances_op.rs
@@ -38,7 +38,7 @@ pub fn handle_expand_instances(binds: &[BufferView]) -> Result<Vec<Vec<u8>>, Com
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{ComputeBackend, Kernel, MockCpu};
+    use crate::{ComputeBackend, Kernel, backend::mock_cpu::MockCpu};
     use std::sync::Arc as StdArc;
 
     #[repr(C)]
