+++ crates/compute/src/kernels/gather_op.rs
@@ -47,7 +47,7 @@ pub fn handle_gather(binds: &[BufferView]) -> Result<Vec<Vec<u8>>, ComputeError>
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{ComputeBackend, Kernel, MockCpu};
+    use crate::{ComputeBackend, Kernel, backend::mock_cpu::MockCpu};
     use std::sync::Arc as StdArc;
 
     #[test]
